userland: remove test-bin in favor of uwutest :^)
