+++ userland/Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["aero_shell", "init", "test-bin"]
+members = ["aero_shell", "init"]
