--- userland/test-bin/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "test-bin"
-version = "0.1.0"
-authors = ["czapek1337 <czapek1337@gmail.com>"]
-edition = "2021"
-
-[dependencies]
-aero_syscall = { path = "../../src/aero_syscall" }
-std = { path = "../aero_std" }
